wip: fix IPV6_PKTINFO ancillary data to override IPV6_PKTINFO socket option
TODO:

assuming this works, it's because sk_bound_dev_if is not set by sticky IPV6_PKTINFO
(it only _checks_ against bound_dev).

this raises the question whether connect() checks against sticky pktinfo,
same as it does for conflict with bound_dev.  looks like it doesn't.
ah, oops, that's just an API consistency issue though.

real question which we haven't tested is whether
you can send from a link-local address on the wrong interface.

It's quite possible there's code that can run that check,
but I haven't been running into it where I've been looking.

however, this is only an RFC violation.
it is not relevant to the security issues I was trying to cover.