KVM: x86: lapic_timer.pending is not a counter (cleanup)
lapic_timer.pending is a flag which is compared to zero, and later cleared.
So we can can set it as a flag, instead of incrementing it as a counter.
Flags are simpler, so this makes the code simpler to understand.

I think it doesn't strictly rely on atomic_set/WRITE_ONCE/volatile
semantics either.  But, they can be useful to indicate a shared-memory
protocol and there's no performance cost.